requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.5
Flask-Cors==4.0.0
python-dotenv==1.0.0
PyJWT==2.8.0
gunicorn==21.2.0
psycopg2-binary